# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.ChangePasswordRequest:
    properties:
      new_password:
        example: NewSecret@123
        minLength: 8
        type: string
      old_password:
        example: secret123
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  controller.ForgotPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
...
  controller.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  controller.LoginResponse:
    properties:
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      role:
        example: student
        type: string
      token:
        example: eyJhbGci...
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  controller.RegisterRequest:
    properties:
      date_of_birth:
        example: "1998-05-20"
        type: string
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      password:
        example: Secret@123
        minLength: 8
        type: string
      phone:
        example: "+919876543210"
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  controller.RegisterResponse:
    properties:
      message:
        example: registration successful
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  controller.ResetPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
      new_password:
        example: NewSecret@123
        minLength: 8
        type: string
      otp:
        example: "042913"
        type: string
    required:
    - email
    - new_password
    - otp
    type: object
...
host: localhost:8080
info:
  contact:
    email: umangagarwal9099@gmail.com
    name: Umang Agarwal
  description: 'Vedex backend API. Authenticate via POST /api/v1/auth/login to get
    a JWT, then pass it as `Authorization: Bearer <token>` on protected endpoints.'
  title: Vedex API
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change the logged-in user's password. Requires the current password.
      parameters:
      - description: Old and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Incorrect current password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a 6-digit OTP to the given email if it belongs to a registered
        account. Always returns 200, whether or not the email is registered, to avoid
        leaking which emails have accounts. The OTP expires after 5 minutes.
      parameters:
      - description: Account email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent if the email is registered
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Forgot password — request OTP
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: 'Single login endpoint for all roles. Returns a JWT — send it as
        `Authorization: Bearer <token>` on protected requests.'
      parameters:
      - description: Email and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.LoginResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user. A unique user ID is generated automatically
        and returned in the response.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.RegisterResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already registered
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Verifies the OTP emailed via /auth/forgot-password and sets a new
        password. The OTP is single-use and expires 5 minutes after being issued.
      parameters:
      - description: Email, OTP, and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid or expired OTP / validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password with OTP
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"